refactor(codegen): decompose x86_64 and AArch64 pipelines into composable passes
x86_64: split monolithic runFunctionPipeline into legalize/allocate/
optimize/emit phases with public module-level APIs (legalizeModuleToMIR,
allocateModuleMIR, emitMIRToAssembly, emitMIRToBinary). New
PeepholePass.cpp/hpp as proper pass. selectTarget() made public.

AArch64: PassManager-based composition replacing direct function calls.
EH-sensitive module bypass for IL optimizations. Scheduler and
BlockLayout passes added to O1+ pipeline. Virtual register space
partitioned (kFirstVirtualRegId, kPhiVRegStart, kCrossBlockSpillKeyStart)
with overflow guards.

Common: CodeSection::appendSection() for section merging with symbol/
reloc rebasing, DebugLineTable::append() with address bias, seedDebugFiles
helper for source-path-aware debug tables, FrameLayout::ensureSpill vreg
widened to uint32_t.

Runtime: rt_trap_get_ip/rt_trap_set_ip for instruction pointer recovery,
rt_err_to_trap_kind error code classification, rt_trap_error_make and
rt_trap_raise_error convenience functions.

Docs: backend.md pipeline diagram updated, native-assembler.md AArch64
pipeline, x86_64.md pass table corrected.

+++ src/codegen/aarch64/FrameBuilder.cpp
@@ -108,7 +108,7 @@ int FrameBuilder::localOffset(unsigned tempId) const {
     return fn_->frame.getLocalOffset(tempId);
 }
 
-int FrameBuilder::ensureSpill(uint16_t vreg, int sizeBytes, int alignBytes) {
+int FrameBuilder::ensureSpill(uint32_t vreg, int sizeBytes, int alignBytes) {
     for (const auto &S : fn_->frame.spills)
         if (S.vreg == vreg)
             return S.offset;
@@ -117,7 +117,7 @@ int FrameBuilder::ensureSpill(uint16_t vreg, int sizeBytes, int alignBytes) {
     return off;
 }
 
-int FrameBuilder::ensureSpillWithReuse(uint16_t vreg,
+int FrameBuilder::ensureSpillWithReuse(uint32_t vreg,
                                        unsigned lastUseInstrIdx,
                                        unsigned currentInstrIdx,
                                        int sizeBytes,
